last legacy function in ariadne files reverted
Still doesn't work. The inputs are the issue and cannot be dealt with by the legacy functions as far as I can
tell. Perhaps this state might deal with more normal levels fine though. Will need to test more normal levels
where the pathfinding bug didn't used to happen.

In the mean time I'll be working on dealing with the extreme edge cases to see if they can be avoided even in
these more extreme levels that cause the immediate hang.