Removed redundant function call; imps get local prettying jobs
The original function call to assign imp jobs based on what is walkable from the dungeon heart turns out to be redundant. Imps get claim and fortify jobs based
on what is accessible to them regardless of where that is, be that isolated from the dungeon but attached to a claimed spot where you dropped them, or somewhere
attached to the dungeon.